+++ apps/dashboard/app/controllers/dashboard_controller.rb
@@ -1,6 +1,10 @@
 class DashboardController < ApplicationController
   def index
-    @motd = MotdFile.new.formatter
+    begin
+      @motd = MotdFile.new.formatter
+    rescue StandardError => e
+      flash.now[:alert] = t('dashboard.motd_erb_render_error', error_message: e.message)
+    end
     set_my_quotas
   end