+++ mod_ood_proxy/lib/node_proxy.lua
@@ -10,9 +10,10 @@ local http     = require 'ood.http'
 --]]
 function node_proxy_handler(r)
   -- read in OOD specific settings defined in Apache config
-  local user_map_cmd = r.subprocess_env['OOD_USER_MAP_CMD']
-  local user_env     = r.subprocess_env['OOD_USER_ENV']
-  local map_fail_uri = r.subprocess_env['OOD_MAP_FAIL_URI']
+  local user_map_match  = r.subprocess_env['OOD_USER_MAP_MATCH']
+  local user_map_cmd    = r.subprocess_env['OOD_USER_MAP_CMD']
+  local user_env        = r.subprocess_env['OOD_USER_ENV']
+  local map_fail_uri    = r.subprocess_env['OOD_MAP_FAIL_URI']
 
   -- read in <LocationMatch> regular expression captures
   local host = r.subprocess_env['MATCH_HOST']
@@ -20,7 +21,7 @@ function node_proxy_handler(r)
   local uri  = r.subprocess_env['MATCH_URI']
 
   -- get the system-level user name
-  local user = user_map.map(r, user_map_cmd, user_env and r.subprocess_env[user_env] or r.user)
+  local user = user_map.map(r, user_map_match, user_map_cmd, user_env and r.subprocess_env[user_env] or r.user)
   if not user then
     if map_fail_uri then
       return http.http302(r, map_fail_uri .. "?redir=" .. r:escape(r.unparsed_uri))