+++ nginx_stage/test/minitest_helper.rb
@@ -1,4 +1,0 @@
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'nginx_stage'
-
-require 'minitest/autorun'